+++ R/import2Dspectra.R
@@ -1,5 +1,4 @@
 #'
-#'
 #' Import 2D Spectroscopic Data
 #' 
 #' This function imports a single file (for instance, a csv) containing a 2D
@@ -16,14 +15,15 @@
 #'
 #' @param ... Parameters to be passed to \code{\link{read.table}}.
 #'
-#' @section ASCII Format Codes:
+#' @section ASCII Format Codes for Data in Three Columns:
 #' ASCII format codes are constructed in two parts separated by a hyphen.  The first part gives
 #' the order of the columns in the file, e.g. F2F1Z means the first column has the F2 values,
 #' the second column has the F1 values and the third the intensities.  The second part of the
 #' format code relates to the order of the rows, i.e. which column varies fastest and in what direction.
 #' These codes are best understood in relation to how the data is stored internally in a matrix.
 #' The internal matrix is organized exactly as the data appears on the screen, with F2 decreasing
-#' left-to-right, and F1 increasing top-to-bottom. There are four possibilities:
+#' left-to-right, and F1 increasing top-to-bottom. There are four possibilities (not all implemented,
+#' please e-mail for assistance):
 #' \itemize{
 #'   \item Data in the file consists of F2 slices starting at the bottom (max F1).  Use \code{F2decF1}.
 #'   \item Data in the file consists of F2 slices starting at the top (min F1).  Use \code{F2incF1}.
@@ -38,6 +38,9 @@
 #'         are taken from the dimension of the matrix.  After import, you will have to manually
 #'         fix the F2 and F1 values.  You may also have to transpose the matrices manually, or
 #'         perhaps invert the order of the rows or columns.
+#'   \item \code{Btotxt}.  This format imports Bruker data written to a file using the Bruker
+#'         "totxt" command.  Tested with TopSpin 4.0.7.  This format is read via \code{readLines}
+#'         and thus \ldots does not apply.
 #' }
 
 #'
@@ -92,5 +95,42 @@ import2Dspectra <- function(file, fmt, nF2, ...) {
     return(ans) 
   } # end of fmt = "F2F1Z-F2decF1"
   
-  if (!valid) stop("fmt not recognized")
+  if (fmt == "Btotxt") {
+  	# intensities (z values) are stored in rows; brief header present
+  	# using readLines here, sep, dec etc don't apply
+  	valid <- TRUE
+    raw <- readLines(file)
+    
+    # get needed header info
+    NROWS <- grep("# NROWS", raw)
+    nF1 <- as.integer(sub("#\\s{1}NROWS\\s{1}=\\s{1}([0-9]+).*", "\\1", raw[NROWS]))
+    NCOLS <- grep("# NCOLS", raw)
+    nF2 <- as.integer(sub("#\\s{1}NCOLS\\s{1}=\\s{1}([0-9]+).*", "\\1", raw[NCOLS]))
+
+    F1LEFT <- grep("F1LEFT", raw) # this line also has F1RIGHT
+    F1left <- as.numeric(sub("#\\s{1}F1LEFT\\s{1}=\\s{1}(-?[0-9]+\\.[0-9]+).*", "\\1", raw[F1LEFT]))
+    F1right <- as.numeric(sub(".*F1RIGHT\\s{1}=\\s{1}(-?[0-9]+\\.[0-9]+).*", "\\1", raw[F1LEFT]))
+	F1 <- sort(seq(F1left, F1right, length.out = nF1))
+	
+    F2LEFT <- grep("F2LEFT", raw)
+    F2left <- as.numeric(sub("#\\s{1}F2LEFT\\s{1}=\\s{1}(-?[0-9]+\\.[0-9]+).*", "\\1", raw[F2LEFT]))
+    F2right <- as.numeric(sub(".*F2RIGHT\\s{1}=\\s{1}(-?[0-9]+\\.[0-9]+).*", "\\1", raw[F2LEFT]))
+	F2 <- sort(seq(F2left, F2right, length.out = nF2))
+	
+    # now process the intensities
+    # rows are numbered from 0 to nF1-1
+    zstart <- grep("#\\s{1}row\\s{1}=\\s{1}0", raw)
+    zvals <- raw[zstart:length(raw)]
+    zvals <- na.omit(gsub("#.*", NA, zvals)) # remove "# row = number" lines
+    zvals <- as.numeric(zvals)
+    if (length(zvals) != nF1*nF2) stop("No. of intensities didn't match dimensions")
+    
+    M <- matrix(zvals, nrow = nF1, ncol = nF2, byrow = TRUE)
+    M <- M[nrow(M):1,] # reflect around horizontal axis as last row was first in file
+
+    ans <- list(M = M, F2 = F2, F1 = F1)
+    return(ans) 
+  } # end of fmt = "Btotxt"
+
+  if (!valid) stop("fmt not recognized. Contact hanson@depauw.edu for help.")
 }